requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.0
gunicorn==21.2.0